Add back pg session store
apparently, render uses multiple concurrent server instances, so if you reconnect with a different instance it will not remember session data. I need to add session info to the database so taht all server instances can access it. I did this before but removed it. i think this is the final issue with authentication. this stuff is very complicated!