Move frontend to backend
I'm trying to set it up so that the server will serve the static site. maybe this will fix the issues with the cookies. I added a build script to build the client and then added a get request to send the static site. I make use of import.meta.env.PROD variable to keep track if static site is built in development or production, because apparently this is something in VITE. I don't really know what I am doing. If so, i can go ahead and remove all the junk I added just before this.